Set `nat` forward_mode for Vagrant
This is neither better nor worse - it's just the setting which
was known to work. It can later be configured to pillar.
For now, because VMs has to be isolated on the net (due to
licenses which bound to IP addresses possibly in use somewhere else),
use NAT to avoid any conflicts with outside world.